Make linkify a helper library
