+++ .rubocop_todo.yml
@@ -122,10 +122,10 @@ Rails/NotNullColumn:
 # Offense count: 9
 Rails/OutputSafety:
   Exclude:
-    - 'app/helpers/application_helper.rb'
     - 'app/helpers/user_mailer_helper.rb'
     - 'lib/rich_text.rb'
-    - 'test/helpers/application_helper_test.rb'
+    - 'lib/linkify.rb'
+    - 'test/lib/linkify_test.rb'
 
 # Offense count: 41
 Rails/SpecificActionNames:
