+++ test/helpers/application_helper_test.rb
@@ -5,32 +5,6 @@ require "test_helper"
 class ApplicationHelperTest < ActionView::TestCase
   attr_accessor :current_user
 
-  def test_linkify
-    %w[http://example.com/test ftp://example.com/test https://example.com/test].each do |link|
-      text = "Test #{link} is <b>made</b> into a link"
-
-      html = linkify(text)
-      assert_predicate html, :html_safe?
-      assert_dom_equal "Test <a href=\"#{link}\" rel=\"nofollow\" dir=\"auto\">#{link}</a> is &lt;b&gt;made&lt;/b&gt; into a link", html
-
-      html = linkify(text.html_safe)
-      assert_predicate html, :html_safe?
-      assert_dom_equal "Test <a href=\"#{link}\" rel=\"nofollow\" dir=\"auto\">#{link}</a> is <b>made</b> into a link", html
-    end
-
-    %w[test@example.com mailto:test@example.com].each do |link|
-      text = "Test #{link} is not <b>made</b> into a link"
-
-      html = linkify(text)
-      assert_predicate html, :html_safe?
-      assert_dom_equal "Test #{link} is not &lt;b&gt;made&lt;/b&gt; into a link", html
-
-      html = linkify(text.html_safe)
-      assert_predicate html, :html_safe?
-      assert_dom_equal "Test #{link} is not <b>made</b> into a link", html
-    end
-  end
-
   def test_rss_link_to
     link = rss_link_to(:controller => :diary_entries, :action => :rss)
     assert_dom_equal "<a href=\"/diary/rss\"><i class=\"bi bi-rss-fill fs-6\" aria-hidden=\"true\"></i></a>", link
